Add files via upload
The following files include two encoders and one decoder. The encoder script encodes a text input using the western musical scale, and the chord-encoder script uses simultaneous notes to make up chords to encode data. The decoder script starts the decoding process upon receiving a "starting signal" that the encoders add to the sound files initially and ends it upon receiving the "Roger signal" that signals the end of an audio message.